+++ source/Nuke.Common/Execution/TargetDefinitionLoader.cs
@@ -6,25 +6,32 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using JetBrains.Annotations;
 using Nuke.Common.Utilities;
+using Nuke.Common.Utilities.Collections;
 
 namespace Nuke.Common.Execution
 {
     internal static class TargetDefinitionLoader
     {
-        public static IReadOnlyCollection<TargetDefinition> GetExecutingTargets(NukeBuild build)
+        public static IReadOnlyCollection<TargetDefinition> GetExecutingTargets(NukeBuild build, [CanBeNull] string[] invokedTargetNames = null)
         {
             ControlFlow.Assert(build.TargetDefinitions.All(x => !x.Name.EqualsOrdinalIgnoreCase(BuildExecutor.DefaultTarget)),
                 $"The name '{BuildExecutor.DefaultTarget}' cannot be used as target name.");
 
-            var invokedTargets = build.InvokedTargets.Select(x => GetDefinition(x, build)).ToList();
+            var invokedTargets = invokedTargetNames?.Select(x => GetDefinition(x, build)).ToList() ?? new List<TargetDefinition>();
             var executingTargets = GetUnfilteredExecutingTargets(build, invokedTargets);
+
             var skippedTargets = executingTargets
                 .Where(x => !invokedTargets.Contains(x) &&
                             build.SkippedTargets != null &&
                             (build.SkippedTargets.Length == 0 ||
                              build.SkippedTargets.Contains(x.Name, StringComparer.OrdinalIgnoreCase))).ToList();
-            skippedTargets.ForEach(x => x.Conditions.Add(() => false));
+            skippedTargets.ForEach(x => x.Skip = true);
+            executingTargets
+                .Where(x => x.DependencyBehavior == DependencyBehavior.Skip)
+                .Where(x => x.Conditions.Any(y => !y()))
+                .ForEach(x => SkipTargetAndDependencies(x, invokedTargets, executingTargets));
 
             string[] GetNames(IEnumerable<TargetDefinition> targets)
                 => targets.Select(x => x.Name).ToArray();
@@ -34,6 +41,35 @@ namespace Nuke.Common.Execution
             ReflectionService.SetValue(build, nameof(NukeBuild.ExecutingTargets), GetNames(executingTargets.Except(skippedTargets)));
 
             return executingTargets;
+        }
+
+        private static void SkipTargetAndDependencies(
+            TargetDefinition targetDefinition,
+            IReadOnlyCollection<TargetDefinition> invokedTargets,
+            IReadOnlyCollection<TargetDefinition> executingTargets,
+            IDictionary<TargetDefinition, List<TargetDefinition>> skipRequestDictionary = null)
+        {
+            skipRequestDictionary = skipRequestDictionary ?? new Dictionary<TargetDefinition, List<TargetDefinition>>();
+            
+            targetDefinition.Skip = true;
+            foreach (var dependency in targetDefinition.TargetDefinitionDependencies)
+            {
+                if (invokedTargets.Contains(dependency))
+                    continue;
+                
+                var skipRequests = skipRequestDictionary.GetValueOrDefault(dependency);
+                if (skipRequests == null)
+                    skipRequests = skipRequestDictionary[dependency] = new List<TargetDefinition>();
+                
+                var executingDependentTargets = executingTargets
+                    .Where(x => x != targetDefinition)
+                    .Where(x => x.TargetDefinitionDependencies.Contains(dependency) && !skipRequests.Contains(x));
+                
+                if (executingDependentTargets.Any())
+                    skipRequests.Add(targetDefinition);
+                else
+                    SkipTargetAndDependencies(dependency, invokedTargets, executingTargets, skipRequestDictionary);
+            }
         }
 
         private static TargetDefinition GetDefinition(