+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [vNext]
+- Added `ITargetDefinition.WhenSkipped` to specify dependency behavior for skipped targets.
 
 ## [0.5.3] / 2018-06-12
 - Fixed global tool to search build scripts also in current directory.